backend/python/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
celery[redis]==5.4.0
redis==5.2.1
boto3==1.36.2
pypdf==5.1.0
pdf2image==1.17.0
Pillow==11.1.0
httpx==0.28.1
structlog==24.4.0
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0
